+++ apps/tergite-mss/CHANGELOG.md
@@ -18,6 +18,7 @@ and this project follows versions of format `{year}.{month}.{patch_number}`.
 - Added back the `/auth/app/{OATH_CLIENT}` URLs for backward compatibility
 - Added filtering bookings at `/bookings/{backend}` endpoint by `user_id`
 - Added sorting bookings at `/bookings/{backend}` endpoint by any field
+- Added the "backend" property to the `BookingsConfig` and the `Booking` schemas returned to the client
 
 ## [2025.09.0] - 2025-10-02